fixes migration and moves function
